move all adapter interface to events
