+++ pkg/adapter/events/configuration.go
@@ -1,0 +1,11 @@
+package events
+
+type ConfigurationCenter interface {
+	Start() error
+
+	Events() <-chan *ConfigEvent
+
+	FindConfiguratorConfig(serviceName string) *ConfiguratorConfig
+
+	Stop() error
+}
