+++ pkg/adapter/events/registry.go
@@ -1,19 +1,15 @@
-package adapter
+package events
 
-import (
-	"github.com/mesh-operator/pkg/adapter/events"
-)
-
-type RegistryClient interface {
+type Registry interface {
 	Start() error
 
 	// Received the notification events indicates services or instances has been modified.
-	Events() <-chan *events.ServiceEvent
+	Events() <-chan *ServiceEvent
 
 	// If you need to make the application name explicit, you can find it with a service name from registry client.
 	// FindAppIdentifier(serviceName string) string
 
-	GetCachedService(serviceName string) *events.Service
+	GetCachedService(serviceName string) *Service
 
 	Stop()
 }
